refactor(core): colocate model call recorder
